+++ packages/yourloops/components/monitoring-alert/monitoring-alerts-card.tsx
@@ -105,7 +105,6 @@ function MonitoringAlertsCard(props: MonitoringAlertsCardProps): JSX.Element {
             id="configure-icon-button-id"
             aria-label={t('configure-monitoring-alerts')}
             data-testid="monitoring-alert-card-configure-button"
-            data-stonlyid="monitoring-alerts-card-configure-button"
             onClick={() => {
               navigate(`..${AppUserRoute.PatientProfile}#${MONITORING_ALERTS_SECTION_ID}`, { relative: 'path' })
             }}